Rename: 클래스 이름 수정, 위치 변경
