+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/notification/domain/MessageSender.java
@@ -1,10 +1,10 @@
-package com.rollingpaper.ggeujeogggeujeog.notification.infrastructure;
+package com.rollingpaper.ggeujeogggeujeog.notification.domain;
 
 import com.rollingpaper.ggeujeogggeujeog.notification.infrastructure.dto.NotificationRequestDto;
 
-public interface PushService {
+public interface MessageSender {
 
-	default void sendMessage(NotificationRequestDto dto) {
+	default void sendNotification(NotificationRequestDto dto) {
 		throw new UnsupportedOperationException("지원하지 않는 메소드입니다.");
 	}
 }
